fix: Update dat description and links
Dat has evolved into a protocol for data sharing
that allows for data tracking, versioning and a p2p
sharing protocol for data, has already some usecases
and community tools and projects.
Some links were also not up to date so, fixed thoses
and included a community for dat projects.